Added more memory information to job dictionary.
Slurm reports memory either as MinMemoryCPU or MinMemoryNode, depending
on how a given job was submitted.  This commit adds keys that shows
whether memory for the job was submitted per cpu or per node.

If the job was submitted per cpu, the new key/value pairs look like:

 u'mem_per_cpu': True,
 u'mem_per_node': False,
 u'min_memory_cpu': 768,
 u'min_memory_node': None,

If the job was submitted per node, the new key/value pairs look like:

 u'mem_per_cpu': False,
 u'mem_per_node': True,
 u'min_memory_cpu': None,
 u'min_memory_node': 16384,

Memory units are in MB.