Modified find_id() method in job class to return list.
Previously, the find_id() method would return the values in dictionary
format, but only return the first item in the list.  This works for
single jobs, but for job arrays, it returns the record for the main
array jobid, but not the subjobs.

Now, find_id() returns a list of dictionaries.  Users will have to
iterate over the list to get the values for each job record returned by
slurm_load_job().