Organize Slurm C-API in a seperate Cython package
This removes the single-file slurm.pxd and instead
uses cythons capability to create cython-packages,
by simple creating a file "__init__.pxd" in a folder,
so that package can also be "cimported" anywhere.

Nothing really changes in the pyslurm.pyx itself,
importing the C-API stays the same. But now it is
possible to organize the C-API a bit better, for example
putting any extra functions which are not part of the
slurm headers directly but found in the libslurm.so into
a different file.

This also directly adds the appropriate header definitions
for Slurm major release 22.05

Header definitions are put into "header.pxi"
Extra libslurm.so are put into "extra.pxi"
Additional helpers are in "helpers.pxi"