Add a script to automatically generate new header bindings
This adds a script pyslurm_bindgen.py to automatically
translate Slurms header files into a single .pxd file
with the help of autopxd2

Additionally, it translates any simple C-macros (#define)
in the headers and makes it available in the .pxd file with
its appropriate data type.
It is mostly useful to upgrade to a new major slurm release.

This makes the jinja2 approach obsolete and it can be deleted.