Use libslurm.so instead of libslurmfull.so
libslurm should be used for interfacing with the C-API,
libslurmfull is more internal to the Slurm tools itself and cannot
be guaranteed to be stable when used externally.

No functions from libslurmfull were actually used
in pyslurm.pyx so we can safely make the switch now.
Also removes a few functions in slurm/extra.pxi, which are
in libslurmfull but not used anywhere in the code

Fixes #209