Backport: Build System and Versioning changes (#239)
* Rework the build system

- Adds a pyproject.toml with build dependencies, so users can also easily
  do a "pip install"

- Update MANIFEST.in file for sdist

- add a test_requirements.txt for easy-install of any dependencies
  necessary for testing

- Update README with build instructions

Changes in setup.py:

- remove Bluegene stuff, it wasn't used anywhere in the code at all
- make some cosmetic changes
- allow doing things like "sdist" and "clean"
  without depending on Cython
- Depend atleast on Cython 0.29.30 as minimum version now.
  This way we can mostly make sure that users have recent Cython
  version to compile

Version naming is changed:

Make up the pyslurm version from the Slurm Major release (e.g. 22.5)
and the current pyslurm patch-level for this major release, so we have
for example:

22.5.0

We must make sure (document it) that users don't confuse this with
Slurms patch version

* Disable the auto_pickle feature which was causing
that pyslurm wasn't able to be compiled on some kernels.

auto pickling may also not be really needed in pyslurm,
because by default classes with pointers/structs as
attributes aren't generated with pickle support by
cython anyway.

For more info, check #236

Fixes #236

* Use libslurm.so instead of libslurmfull.so

libslurm should be used for interfacing with the C-API,
libslurmfull is more internal to the Slurm tools itself and cannot
be guaranteed to be stable when used externally.

No functions from libslurmfull were actually used
in pyslurm.pyx so we can safely make the switch now.
Also removes a few functions in slurm/extra.pxi, which are
in libslurmfull but not used anywhere in the code

Fixes #209

Co-authored-by: tazend <toni.harzendorf@gmail.com>