Add slurmctld functions and full slurm.conf access (#367)
Further changes:

* move cpu_freq_int_to_str to utils.helpers

* cstr: add configurable delim to to_list

* update partition unit tests

* tests: integration tests for slurmctld

* partition: use _get_memory from slurmctld module

* mkdocs.yml: add pymdownx.magiclink

* docs: add griffe_exts.py script for prettier docs

This allows to dynamically link to the Slurm Documentation with
different Slurm versions.

For example it will replace this in the docstring:

{slurm.conf#OPT_MpiParams}

with this in the finally rendered docs, when we are on pyslurm 24.11.x:

https://slurm.schedmd.com/archive/slurm-24.11-latest/slurm.conf.html#OPT_MpiParams

This way we can dynamically link to the Slurm docs.

* docs: add slurmctld reference API page

* docs: remove old config class documentation

* update common tests

* remove the setup.py clean step from builddocs.sh

(cherry picked from commit ff3dc424cc84e8390f6a0027148c946c3c33536d)